Changes from https://github.com/kaldi-asr/kaldi/commit/5d2cc3#commitcomment-11321531
